requirements.txt
awsebcli==3.12.3
blessed==1.14.2
botocore==1.8.48
cement==2.8.2
click==6.7
colorama==0.3.7
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
lxml==4.1.1
MarkupSafe==1.0
numpy==1.14.1
pandas==0.22.0
pathspec==0.5.5
PyMySQL==0.8.0
python-amazon-product-api==0.2.8
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
requests==2.9.1
semantic-version==2.5.0
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
titlecase==0.12.0
wcwidth==0.1.7
websocket-client==0.47.0
Werkzeug==0.14.1
